net: ipc_router: Register services only on client port
Allowing services to be registered on a non-client port will cause either
an existing service or a control port to be over-written. This will cause
undefined functional behavior.

Allow the services to be registered only on client ports.

FPIIM-388
FPII-2829: Elevation of privilege vulnerability in Qualcomm networking driver (device specific) CVE-2017-0463 A-33277611 QC-CR#1101792

Change-Id: I17f2e3473c1c3c69c544b5cf8f87ea413fd9cf9a